+++ src/containers/forms/ContactForm.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { navigate } from 'gatsby'
 
 import { ContactForm as ContactFormComponent } from '../../components/forms'
-import { FormHandler } from '../../services'
+import FormHandler from './FormHandler'
 
 /**
  * A contact form container that controls submission of the contact form, displays toasts and navigates to '/' on
