+++ src/containers/forms/FormHandler.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { axios } from '../../axios'
+import { axios } from '../../../axios'
 
 /**
  * A Form Handler service that renders a react component containing a form, handling submission to the
