+++ src/containers/forms/index.js
@@ -1,2 +1,3 @@
 export { default as ContactForm } from './ContactForm'
+export { default as FormHandler } from './FormHandler'
 export { default as SubscribeForm } from './SubscribeForm'
