+++ src/containers/index.js
@@ -1,2 +1,2 @@
-export { ContactForm, SubscribeForm } from './forms'
+export { ContactForm, FormHandler, SubscribeForm } from './forms'
 export { CallToAction, setCtaOpen, setCtaClosed } from './elements'
